Create a tiny pubsub library
It turns out that nsIObserver is a global pubsub mechanism (global
across all Firefox windows), so it's not a natural fit for messaging
between components attached to a single window. This little bit of
pubsub code seemed simpler than trying to append window identifiers to
the messages and implement some sort of wrapper.

Closes #36.