go back to verifying we're at the zeroth item before stealing focus
although really, when would you ever get new results when the user was
focused in the popup? you get results when the user is typing in the
urlbar. *maybe* you steal focus with someone who is super duper quick at
down-arrowing into the popup, but is that worth slowing things down
further?