EOD Status update:
Working on debugging the `TypeError: rows[(numRows - 1)] is undefined`
error, I added some watch expressions to the Browser Toolbox, and it
looks like maxRows is always set to -1:

![](https://www.dropbox.com/s/albiwwzpvtfwr68/Screen%20Shot%202016-08-25%20at%205.12.10%20PM.png?dl=0&raw=true)

I had thought that the problem was the JS popuphiding listener
clobbering the existing XBL popuphiding listener (which is supposed to
reset maxRows to its usual value (this code can be found
[in
dxr](https://dxr.mozilla.org/mozilla-central/source/toolkit/content/widgets/autocomplete.xml#990-996)):

```js
  // Reset the maxRows property to the cached "normal" value, and reset
  // _normalMaxRows so that we can detect whether it was set by the input
  // when the popupshowing handler runs.

  // Null-check this.mInput; see bug 1017914
  if (this.mInput)
    this.mInput.maxRows = this._normalMaxRows;
  this._normalMaxRows = -1;
```

That line where `this.mInput.maxRows` gets reset to
`this._normalMaxRows` should reset maxRows to 6 or 10. It isn't. This
might be because _normalMaxRows is never properly updated, or it might
be because this.mInput is falsy. Need to explore.

I changed our popuphiding listener to a popuphidden listener a few
commits ago, but it didn't fix the bug. It did fix another sub-bug
(instead of the popup rendering empty, it renders the stale results
list), so I guess that's some kind of progress, but still doesn't fix
the popup after the first round of results is rendered.

Couple options to explore:
- Figure out if we can just override either maxRows or _normalMaxRows
  with a no-op setter that always forces it to be a positive number.
	Some comments in the code suggest that -1 is a special value that's
	used to check the popup state, need to dig those up - autocomplete.xml
	I think?
- Hacky, but would work: copy-paste the code we want into a handler that
	overrides the old handler
- Super hacky: listen for and undo changes to maxRows/normalMaxRows,
  either via polling or, if possible, subscribing to change events on
  the attribute. Need to look up if it's field or property XBL elements
  that fire change events when the setter is invoked.

I really like this way of keeping the WIP notes with the WIP code :tada: