What if we override invalidate() to set maxRows?
invalidate() calls _invalidate() (which does all the work), recall this
is because XBL inheritance doesn't allow superclass methods to be
called.

Because invalidate() just wraps the function that does the work, it
contains hardly any content:

inside autocomplete.xml:
```js
  <method name="invalidate">
    <parameter name="reason"/>
    <body>
      <![CDATA[
      // Don't bother doing work if we're not even showing
      if (!this.mPopupOpen)
        return;

      this._invalidate(reason);
      ]]>
    </body>
  </method>
```

and urlbarBindings.xml doesn't override invalidate at all, just calls
this.invalidate().

So, we could override invalidate to duplicate the "if not open return"
logic, but also ensure maxRows has a sane value.

This might fix the maxrows - 1 bug and we could call it a day.

Let's give it a shot.